Minor chagnes

+++ Bookings/BookingApp/RequestView.py
@@ -11,6 +11,7 @@ from rest_framework.response import Response
 
 from .models import *
 from .serializers import *
+from django.db.models import Q
 
 
 class RequestListView(generics.ListCreateAPIView):
@@ -33,7 +34,7 @@ class RequestListView(generics.ListCreateAPIView):
             team=get_object_or_404(Team.objects.all(),teamId=team_id)
             event=get_object_or_404(Event.objects.all(),eventId=event_id)
             query=queryset.filter(teamId=team).filter(event=event).all()
-            print(query)
+            
             if query :
                 return Response({"message": "Request already pending"})
             request = Request(
@@ -63,16 +64,26 @@ class RequestListView(generics.ListCreateAPIView):
             timeStart = datetime.strptime(time, date_time_format).time()
           
             for indv in users:
-                print(indv)
+                
                 userind=get_object_or_404(User.objects.all(),userId=indv)
                
                 userobj.append(userind)
                
-                query=queryset.filter(  individuals__in=userobj,    amenity=amenity,dateSlot=dateSlot,timeStart__contains=timeStart, ).all()
+                query=queryset.filter(  individuals__in=userobj,amenity=amenity,dateSlot=dateSlot,timeStart__contains=timeStart, ).all()
                 
                 if query.count()!=0:
                     return Response({'message':f"Request Already Pending for {userind.userName}"})
-           
+
+                query=Booking.objects.all().filter(  individuals__in=userobj,amenity=amenity,dateSlot=dateSlot,timeStart__contains=timeStart, ).all()
+                
+                if query.count()!=0:
+                    return Response({'message':f"Booking Already Pending for {userind.userName}"})
+            
+            availability=AmenitySlot.objects.all().filter(amenity=amenity,amenityDate=dateSlot,amenitySlotStart__contains=timeStart)
+            if (availability.count()>0 and availability.first().capacity<len(users)) or (amenity.capacity<len(users)):
+                return Response({"message":"Count of people exceeds capacity"})
+            
+
             request = Request(
                
                 amenity=amenity,
@@ -87,7 +98,7 @@ class RequestListView(generics.ListCreateAPIView):
             
             request.save()
             request.individuals.set(userobj)
-        return Response({"message": "Request created successfully"}, status=status.HTTP_201_CREATED)
+        return Response({"message": "Request created successfully","code":200}, status=status.HTTP_201_CREATED)
 
 
 class RequestDetailView(generics.RetrieveUpdateDestroyAPIView):
@@ -100,6 +111,13 @@ class RequestProvider(generics.ListAPIView):
     def get_queryset(self):
         userId = self.kwargs['userProvider']
         queryset = Request.objects.filter(userProvider=userId)
+        return queryset
+class RequestByUser(generics.ListAPIView):
+    serializer_class=RequestSerializer
+    def get_queryset(self):
+        userId = self.kwargs['userId']
+
+        queryset = Request.objects.filter(   Q(individuals=userId) | Q(teamId__users=userId))
         return queryset
 class RequestToBooking(generics.RetrieveUpdateDestroyAPIView):
     queryset=Request.objects.all()
@@ -122,14 +140,45 @@ class RequestToBooking(generics.RetrieveUpdateDestroyAPIView):
     timeStart=requestObj.timeStart,
             verified=False
         )
-        booking.save()   
-        for individual in requestObj.individuals.all():
-                booking.individuals.add(individual)  
-        requestObj.delete()    
-        if booking.event is not None and booking.teamId is not None:
-            team = Team.objects.get(pk=booking.teamId)
-            team.bookedEvents.add(booking.event)
-            team.save()   
+        if booking.amenity is not None:
+            availability=AmenitySlot.objects.all().filter(amenity=booking.amenity,amenityDate=booking.dateSlot,amenitySlotStart__contains=booking.timeStart)
+            if (availability.count()==0):
+                if (booking.amenity.capacity-len(requestObj.individuals.all())>=0):
+                    amenitySlot=AmenitySlot(
+                        amenity=booking.amenity,
+        amenityDate=booking.dateSlot,
+        amenitySlotStart=booking.timeStart,
+        amenitySlotEnd=AmenitySlot.objects.all().filter(amenity=booking.amenity,amenityDate=None,amenitySlotStart__contains=booking.timeStart).first().amenitySlotEnd,
+        
+        
+        capacity=booking.amenity.capacity-len(requestObj.individuals.all())
+                )
+                    amenitySlot.save()
+                    booking.save()
+                    for individual in requestObj.individuals.all():
+                            booking.individuals.add(individual)  
+                    requestObj.delete()    
+                    if booking.event is not None and booking.teamId is not None:
+                        team = Team.objects.get(pk=booking.teamId)
+                        team.bookedEvents.add(booking.event)
+                        team.save()  
+            else:
+                availability=availability.first()
+                if (availability.capacity-len(requestObj.individuals.all())>=0):
+                    availability.capacity-=len((requestObj.individuals.all()))
+                    availability.save()
+                    booking.save()
+
+            
+                    for individual in requestObj.individuals.all():
+                            booking.individuals.add(individual)  
+                    requestObj.delete()    
+                    if booking.event is not None and booking.teamId is not None:
+                        team = Team.objects.get(pk=booking.teamId)
+                        team.bookedEvents.add(booking.event)
+                        team.save()   
+        
+
         return Response( status=status.HTTP_201_CREATED)
